+++ mpc_energy/CHANGELOG.md
@@ -3,6 +3,7 @@
 * Fixed issue causing battery discharge cost input not adjusting MPC behaviour
 * Added price forecast uncertainty consideration in MPC plan.
 * Added configurable log level
+* Increased Amber forecast horizon to use as much as the API returns (can vary between 12-48 hours).
 
 ## 0.4.0
 * Added Solar Curtailment Sensor to show when the system is likely curtailing and why